Move Tutorial Geometry Done
I finished the geometry of the movement tutorial and added a small challenge at the end with a button and a door. The button has a script that resets the position of the door when e is pressed within range of the button and updates new UI text called "Display Text". The door has a script on it that forces it to go down in the Y direction until a certain y value is reached.

Additionally, I edited the lighting slightly to increase ambient light. I think I need to decrease global fog density but will do that later.