Fixed the sprint jumping issue
Made it so sprint jumping costs stamina the entire time you are jumping.

Modified the if statement for sprinting. Now also checks to see if the magnitude of the velocity of the character is greater that 0.3f. This is to ensure that the player only uses stamina if they are moving.